backend/backend/src/main/java/CS3300/service/ClassService.java
package CS3300.service;

import CS3300.repository.ClassRepository;
import CS3300.schema.Class;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class ClassService {

    @Autowired
    ClassRepository classRepository;

    public List<Class> getAllClasses() {
        return classRepository.findAll();
    }

    public Optional<Class> getClassById(Long id) {
        return classRepository.findById(id);
    }

    public List<Class> getClassesByOwnerId(String ownerId) {
        return classRepository.findByOwnerId(ownerId);
    }

    public Class saveClass(Class classEntity) {
        return classRepository.save(classEntity);
    }

    public void deleteClass(Long id) {
        classRepository.deleteById(id);
    }
}